Renaming DisallowComponentCssCounter
